ForeAiBackend/app/pyproject.toml
[project]
name = "foreaibackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4",
    "chromadb",
    "fastapi[standard]>=0.115.11",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "regex>=2024.11.6",
    "requests>=2.32.3",
    "sqlmodel>=0.0.24",
    "uuid>=1.30",
]
